lab-2: switch to doubles

+++ lab-2/include/matrix_mult.hpp
@@ -2,7 +2,7 @@
 
 #include <vector>
 
-using Matrix = std::vector<std::vector<int>>;
+using Matrix = std::vector<std::vector<double>>;
 
 #define LOOP3(a, b, c, BODY) \
     for (int a = 0; a < N; a++) \
